과목 이름 수정 코드 완료
